Create Code2.py
Replace "filename.txt" with the name of the file you want to read in.

The script reads in the file using the with open() statement, which automatically closes the file when the block of code is finished executing. It then reads all lines of the file using readlines() and joins them together into one string using "".join().

Next, the script splits the text into words using the split() method, which splits the string into a list of words based on the whitespace (spaces, tabs, and newlines) between them.

Finally, the script prints out all the words separated by spaces using the join() method, which concatenates all the elements of a list together into a single string using the specified delimiter (in this case, a space).